feat: merge ping and uptime cogs
