+++ bot/exts/evergreen/ping.py
@@ -1,11 +1,15 @@
+from dateutil.relativedelta import relativedelta
+
+import arrow
 from discord import Embed
 from discord.ext import commands
 
+from bot import start_time
 from bot.constants import Colours
 
 
 class Ping(commands.Cog):
-    """Ping the bot to see its latency and state."""
+    """Get info about the bot's ping and uptime."""
 
     def __init__(self, bot: commands.Bot):
         self.bot = bot
@@ -21,7 +25,20 @@ class Ping(commands.Cog):
 
         await ctx.send(embed=embed)
 
+    @commands.command(name="uptime")
+    async def uptime(self, ctx: commands.Context) -> None:
+        """Responds with the uptime of the bot."""
+        difference = relativedelta(start_time - arrow.utcnow())
+        uptime_string = start_time.shift(
+            seconds=-difference.seconds,
+            minutes=-difference.minutes,
+            hours=-difference.hours,
+            days=-difference.days
+        ).humanize()
+
+        await ctx.send(f"I started up {uptime_string}.")
+
 
 def setup(bot: commands.Bot) -> None:
-    """Cog load."""
+    """Load the Ping cog."""
     bot.add_cog(Ping(bot))
